% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helpers.R
\name{censor}
\alias{censor}
\title{Censor dataset}
\usage{
censor(x, p)
}
\arguments{
\item{x}{Numeric \nxd matrix.}

\item{p}{Numeric vector with \code{d} elements.}
}
\value{
Numeric \nxd matrix.
}
\description{
Censors each row of matrix \code{x} with vector \code{p}.
}
\keyword{internal}
