% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{check_pos_def}
\alias{check_pos_def}
\title{Check if a square, symmetric matrix is positive definite}
\usage{
check_pos_def(M, alert = get_alert_function())
}
\description{
Check if a square, symmetric matrix is positive definite
}
