% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{check_split_by_sep}
\alias{check_split_by_sep}
\title{Identify pairs of vertices that are split by a separator}
\usage{
check_split_by_sep(graph, sep, edgeMat)
}
\arguments{
\item{graph}{A graph}

\item{sep}{A set of vertex ids that are used to split the graph}

\item{edgeMat}{A two-column matrix, containing the vertex-paris to be checked}
}
\value{
A logical vector, indicating for each edge whether it is split by \code{sep}
}
\description{
Identify pairs of vertices that are split by a separator
}
