% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations2.R
\name{checkGamma}
\alias{checkGamma}
\alias{checkTheta}
\alias{checkSigma}
\title{Parameter matrix checks}
\usage{
checkGamma(Gamma)

checkTheta(Theta, k, full)

checkSigma(Sigma, k, full)
}
\arguments{
\item{Gamma}{Numeric \dxd variogram matrix.}

\item{Theta}{Numeric \dxd or \d1xd1 precision matrix.}

\item{k}{\code{NULL} if the input/output matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{full}{Logical. If \code{TRUE} and \code{!is.null(k)},
the input/output matrix is a \dxd matrix with the kth row filled with zeros.}

\item{Sigma}{Numeric \dxd or \d1xd1 covariance matrix.}
}
\value{
The input matrix, passed through \code{\link{ensure_matrix_symmetry_and_truncate_zeros}},
if it is ok, or throws an error, otherwise.
}
\description{
Checks wheter the matrix given is a valid Huesler-Reiss parameter matrix
in the role of \eGamma, \eTheta, or \eSigma, respectively.
}
