% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{chi.est}
\alias{chi.est}
\title{Estimate \eqn{\chi}}
\usage{
chi.est(data, u, pot = FALSE)
}
\arguments{
\item{data}{Numeric matrix \eqn{n \times 2}{n x 2}. A data matrix
containing two variables.}

\item{u}{Numeric, between 0 and 1. It is the probability threshold used to
compute the \eqn{\chi} coefficient.}

\item{pot}{Boolean. if TRUE, then pot-type estimation of EC is used.
By default, \code{pot = FALSE}.}
}
\value{
Numeric. The empirical \eqn{\chi} coefficient between the 2 variables
in \code{data}.
}
\description{
Estimates the chi coefficient empirically.
}
\details{
The \eqn{\chi} coefficient is a scalar coefficient that represents
the extremal correlation between two variables.
}
