% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Chi2Gamma}
\alias{Chi2Gamma}
\title{Convert \eqn{\chi} to \eqn{\Gamma}}
\usage{
Chi2Gamma(chi)
}
\arguments{
\item{chi}{Numeric between 0 and 1. The extremal correlation coefficient.}
}
\value{
Numeric
}
\description{
Transform the \code{chi} extremal correlation into HR parameter.
}
