% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions_main.R
\name{complete_Gamma_decomposable}
\alias{complete_Gamma_decomposable}
\title{Completion of decomposable Gamma matrices}
\usage{
complete_Gamma_decomposable(Gamma, graph = NULL)
}
\arguments{
\item{Gamma}{A variogram matrix that is specified on the edges of \code{graph}
and the diagonals. All other entries are ignored.}

\item{graph}{A decomposable \link[igraph:make_graph]{igraph::graph} object}
}
\value{
A complete variogram matrix that agrees with \code{Gamma} on the entries
corresponding to edges in \code{graph} and the diagonals.
The corresponding \eqn{\Theta} matrix pdocued by \link{Gamma2Theta} has zeros
in the remaining entries.
}
\description{
Given a decomposable \code{graph} and incomplete variogram matrix \code{Gamma},
returns the full \code{Gamma} matrix implied by the conditional independencies.
}
\seealso{
Other Matrix completions: 
\code{\link{DEPRECATED_complete_Gamma_general}()},
\code{\link{complete_Gamma_general_demo}()},
\code{\link{complete_Gamma_one_step}()},
\code{\link{complete_Gamma}()}
}
\concept{Matrix completions}
