% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions.R
\name{complete_Gamma_general_demo}
\alias{complete_Gamma_general_demo}
\title{Completion of non-decomposable Gamma matrices (demo-version)}
\usage{
complete_Gamma_general_demo(
  Gamma,
  graph = NULL,
  N = 1000,
  tol = 0,
  gList = NULL
)
}
\arguments{
\item{Gamma}{A complete variogram matrix (without any graphical structure)}

\item{graph}{An \link[igraph:make_graph]{igraph::graph} object}

\item{N}{The maximal number of iterations of the algorithm}

\item{tol}{The tolerance to use when checking for zero entries in \code{Theta}}
}
\value{
A nested list, containing the following details: TODO!!!

The corresponding \eqn{\Theta} matrix produced by \link{Gamma2Theta} has values
close to zero in the remaining entries (how close depends on the input
and the number of iterations).
}
\description{
Given a \code{graph} and variogram matrix \code{Gamma}, returns the full \code{Gamma}
matrix implied by the conditional independencies.
This function uses a convergent iterative algorithm.
DEMO VERSION: Returns a lot of details and allows specifying the graph list
that is used. Is way slower than \code{complete_Gamma_general}.
}
\seealso{
Other Matrix completions: 
\code{\link{DEPRECATED_complete_Gamma_general}()},
\code{\link{complete_Gamma_decomposable}()},
\code{\link{complete_Gamma_one_step}()},
\code{\link{complete_Gamma}()}
}
\concept{Matrix completions}
