% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions_main.R
\name{complete_Gamma_one_step}
\alias{complete_Gamma_one_step}
\title{Completion of two-clique decomposable Gamma matrices}
\usage{
complete_Gamma_one_step(Gamma, nA, nC, nB)
}
\description{
Given a decomposable \code{graph} consisting of two cliques and incomplete
variogram matrix \code{Gamma}, returns the full \code{Gamma} matrix implied by the
conditional independencies. The rows/columns of \code{Gamma} must be ordered
such that the clique of size \code{nA} (excluding separator) comes first, then
the separator of size \code{nC}, and then the remaining \code{nB} vertices.
}
\keyword{internal}
