% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_plotting.R
\name{computeLimits}
\alias{computeLimits}
\title{Helper function to compute the axis limits of a plot
with given x, y data and optionally a fixed x-y-ratio and
correcting the latitude/longitude scale at different latitudes}
\usage{
computeLimits(xData, yData, xyRatio = 1, convertLatLong = TRUE, stretch = 1)
}
\description{
Helper function to compute the axis limits of a plot
with given x, y data and optionally a fixed x-y-ratio and
correcting the latitude/longitude scale at different latitudes
}
\keyword{internal}
