% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{data2mpareto}
\alias{data2mpareto}
\title{Convert \code{data} to Pareto scale}
\usage{
data2mpareto(data, p)
}
\arguments{
\item{data}{Numeric matrix of size \eqn{n\times d}{n x d}.}

\item{p}{Numeric between 0 and 1. Probability used for the quantile to
threshold the data.}
}
\value{
Numeric matrix \eqn{m \times d}{m x d}, where \eqn{m} is the number
of rows in the original \code{data} matrix that are above the threshold.
}
\description{
Transfroms the data empirically to multivariate Pareto scale.
}
