% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{DEPRECATED_complete_Gamma_general}
\alias{DEPRECATED_complete_Gamma_general}
\title{DEPRECATED: Completion of non-decomposable Gamma matrices}
\usage{
DEPRECATED_complete_Gamma_general(
  Gamma,
  graph,
  N = 1000,
  tol = 0,
  check_tol = 100
)
}
\arguments{
\item{Gamma}{A complete variogram matrix (without any graphical structure)}

\item{graph}{An \link[igraph:make_graph]{igraph::graph} object}

\item{N}{The maximal number of iterations of the algorithm}

\item{tol}{The tolerance to use when checking for zero entries in \code{Theta}}

\item{check_tol}{After how many iterations to check the tolerance in \code{Theta}}
}
\value{
A matrix that agrees with \code{Gamma} on the entries corresponding to
edges in \code{graph} and the diagonals.
The corresponding \eTheta matrix produced by \link{Gamma2Theta} has values
close to zero in the remaining entries (how close depends on the input
and the number of iterations).
}
\description{
Given a \code{graph} and variogram matrix \code{Gamma}, returns the full \code{Gamma}
matrix implied by the conditional independencies.
This function uses a convergent iterative algorithm.
}
\seealso{
Other Matrix completions: 
\code{\link{complete_Gamma_decomposable}()},
\code{\link{complete_Gamma_general_demo}()},
\code{\link{complete_Gamma_one_step}()},
\code{\link{complete_Gamma}()}
}
\concept{Matrix completions}
