% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{dim_Gamma}
\alias{dim_Gamma}
\title{Is Gamma square matrix?}
\usage{
dim_Gamma(Gamma)
}
\arguments{
\item{Gamma}{Numeric matrix. Matrix representing the variogram of an HR
distribution.}
}
\value{
Numeric. The dimension of the matrix (number of rows and columns, if
the matrix is symmetric). Else, raises an error.
}
\description{
Check if Gamma matrix is square matrix. If so, return the dimension. Else,
raise an error.
}
\seealso{
Other Input checks: 
\code{\link{check_Gamma_and_graph}()},
\code{\link{check_graph}()}
}
\concept{Input checks}
\keyword{internal}
