% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{eglasso}
\alias{eglasso}
\title{Fitting extremal graphical lasso}
\usage{
eglasso(
  Gamma,
  rholist = c(0.1, 0.15, 0.19, 0.205),
  reg_method = c("mb", "glasso"),
  eps = 0.5
)
}
\arguments{
\item{Gamma}{Numeric matrix \eqn{n\times d}{n x d}.
It represents a variogram matrix \eqn{\Gamma}.}

\item{rholist}{Numeric vector of non-negative regularization parameters
for the lasso. For details see \link[glasso:glassopath]{glasso::glassopath}.}

\item{reg_method}{One of \code{"mb"} and \code{"glasso"}.
Default is \code{reg_method = "mb"}.}

\item{eps}{Regularizatoin parameter for covariance matrix.}
}
\value{
List with as many elements as entries in \code{rholist}. Each element of
the list contains:
\describe{
\item{\code{rho}}{The penalty coefficient.}
\item{\code{graph}}{An \link[igraph:make_graph]{igraph::graph} object representing the fitted graph.}
\item{\code{Gamma}}{A numeric \eqn{d\times d}{d x d} estimated variogram matrix \eqn{\Gamma}
corresponding to the fitted graph.}
}
}
\description{
Fits an extremal minimum spanning tree
}
