% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ensure_symmetry}
\alias{ensure_symmetry}
\title{Ensure matrix symmetry}
\usage{
ensure_symmetry(M, tol = Inf)
}
\arguments{
\item{M}{Numeric square matrix.}

\item{tol}{Positive scalar. If the maximum absolute difference between \code{M}
and \code{t(M)} is larger, show a warning.}
}
\value{
The value of \code{(M+t(M))/2}.
}
\description{
Ensures the symmetry of a square matrix by averaging it with its transpose.
Optionally verifies that the matrix was close to symmetric before.
}
