% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{est_vario}
\alias{est_vario}
\title{Estimate \eqn{\Gamma}}
\usage{
est_vario(data, k = NULL, p = NULL)
}
\arguments{
\item{data}{Numeric matrix \eqn{n\times d}{n x d}. Data matrix of
observations following a Huesler-Reiss distribution.}

\item{k}{Integer between 1 and \eqn{d}. Component of the multivariate
observations that is conditioned to be larger than the threshold \code{p}.}

\item{p}{Numeric between 0 and 1. Probability threshold for the
the components \code{k}. If \code{NULL},
it is assumed that \code{data} has already multivariate Pareto margins.}
}
\value{
Numeric matrix \eqn{d \times d}{d x d}. The estimated
variogram of the Huesler-Reiss distribution.
}
\description{
Estimates the variogram of the Huesler-Reiss distribution empirically.
}
