% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helpers.R
\name{fast_diag}
\alias{fast_diag}
\title{Fast computation of diag(y \%\emph{\% M \%}\% t(y))}
\usage{
fast_diag(y, M)
}
\arguments{
\item{y}{Numeric matrix}

\item{M}{Numeric matrix}
}
\value{
Numeric vector
}
\description{
Fast computation of diag(y \%\emph{\% M \%}\% t(y))
}
