% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_param_MLE.R
\name{fillFixedParams}
\alias{fillFixedParams}
\title{Helper function to combine par with fixed params (in init)}
\usage{
fillFixedParams(par, init, fixParams)
}
\description{
Helper function to combine par with fixed params (in init)
}
