% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{findVsep}
\alias{findVsep}
\title{Find a separator set for two vertices}
\usage{
findVsep(graph, v0, v1)
}
\description{
Finds a reasonably small set of vertices that separate \code{v0} and \code{v1} in \code{graph}.
}
\keyword{internal}
