% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions_2.R
\name{fit_graph}
\alias{fit_graph}
\title{Experimental: Fit graph using empirical Theta matrix}
\usage{
fit_graph(data, m)
}
\description{
Experimental: Fit graph using empirical Theta matrix
}
