% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fitInInterval}
\alias{fitInInterval}
\title{Fit value(s) in interval}
\usage{
fitInInterval(x, xMin = -Inf, xMax = Inf)
}
\arguments{
\item{x}{Numeric vector}

\item{xMin}{Numeric vector}

\item{xMax}{Numeric vector}
}
\value{
Numeric vector
}
\description{
Fit value(s) in interval, all arguments are recycled where necessary.
}
