% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights_old}
\alias{flights_old}
\title{DEPRECATED: Flights delay data}
\format{
An object of class \code{list} of length 3.
}
\source{
U.S. Department of Transportation's (DOT) Bureau of Transportation Statistics \url{https://www.bts.gov/}.
}
\usage{
flights_old
}
\description{
A dataset containing daily total delays (Jan 1, 2015 -- Dec 31, 2015)
of Southwest Airlines at different airports in California, Nevada,
Arizona, Utah, and Texas.
}
\details{
This dataset was called \code{flights} in earlier development versions of this package
and is is superseded by the new dataset now called \code{flights}.
}
\keyword{datasets}
