% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{fmpareto_graph_HR}
\alias{fmpareto_graph_HR}
\title{Parameter fitting for multivariate Huesler--Reiss Pareto distributions on block graphs}
\usage{
fmpareto_graph_HR(
  data,
  graph,
  p = NULL,
  method = c("ML", "vario"),
  cens = FALSE
)
}
\arguments{
\item{data}{Numeric matrix of size \eqn{n\times d}{n x d}, where \eqn{n} is the
number of observations and \eqn{d} is the dimension.}

\item{graph}{Undirected graph object from \code{igraph} package with \code{d} vertices.}

\item{p}{Numeric between 0 and 1 or \code{NULL}. If \code{NULL} (default),
it is assumed that the \code{data} are already on multivariate Pareto scale. Else,
\code{p} is used as the probability in the function \code{\link{data2mpareto}}
to standardize the \code{data}.}

\item{method}{One of \verb{"ML", "vario"}.
Default is \code{method = "ML"}.}

\item{cens}{Logical. If true, then censored likelihood contributions are used for
components below the threshold. By default, \code{cens = FALSE}.}
}
\value{
List consisting of:
\item{\code{graph}}{An \code{\link[igraph:make_graph]{igraph::graph()}} object.}
\item{\code{Gamma}}{Numeric \eqn{d\times d}{d x d} estimated variogram matrix \eqn{\Gamma}.}
}
\description{
Fits the parameters of a multivariate Huesler--Reiss Pareto distribution using (censored) likelihood estimation.
See \insertCite{eng2019;textual}{graphicalExtremes} for details.
}
\examples{
## Fitting a 4-dimensional HR distribution

my_graph <- igraph::graph_from_adjacency_matrix(
  rbind(
    c(0, 1, 0, 0),
    c(1, 0, 1, 1),
    c(0, 1, 0, 0),
    c(0, 1, 0, 0)
  ),
  mode = "undirected"
)
n <- 100
Gamma_vec <- c(.5, 1.4, .8)
complete_Gamma(Gamma = Gamma_vec, graph = my_graph) ## full Gamma matrix
edges_to_add <- rbind(c(1, 3), c(1, 4), c(3, 4))

set.seed(123)
my_data <- rmpareto_tree(n, "HR", tree = my_graph, par = Gamma_vec)
my_fit <- fmpareto_graph_HR(my_data,
  graph = my_graph, method = "ML",
  p = NULL, cens = FALSE
)
}
\references{
\insertAllCited{}
}
