% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions_2.R
\name{fmpareto_graph_HR_2}
\alias{fmpareto_graph_HR_2}
\title{Parameter fitting for multivariate Huesler--Reiss Pareto distributions on decomposable graphs}
\usage{
fmpareto_graph_HR_2(data, graph, p = NULL, cens = FALSE)
}
\description{
Similar to \code{\link[=fmpareto_graph_HR]{fmpareto_graph_HR()}}. Differences are:
\itemize{
\item Works with any decomposable graph, not just block graphs
\item Does not support \code{edges_to_add}
}
}
