% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_parameters.R
\name{fmpareto_graph_HR_decomposable}
\alias{fmpareto_graph_HR_decomposable}
\title{Parameter fitting for multivariate Huesler--Reiss Pareto distributions on decomposable graphs}
\usage{
fmpareto_graph_HR_decomposable(data, graph, p = NULL, cens = FALSE)
}
\arguments{
\item{data}{Numeric matrix of size \eqn{n\times d}{n x d}, where \eqn{n} is the
number of observations and \eqn{d} is the dimension.}

\item{graph}{Decomposable graph object from \code{igraph} package.}

\item{p}{Numeric between 0 and 1 or \code{NULL}. If \code{NULL} (default),
it is assumed that the \code{data} are already on multivariate Pareto scale. Else,
\code{p} is used as the probability in the function \link{data2mpareto}
to standardize the \code{data}.}

\item{cens}{Logical. If true, then censored likelihood contributions are used for
components below the threshold. By default, \code{cens = FALSE}.}
}
\description{
Similar to \code{\link[=fmpareto_graph_HR]{fmpareto_graph_HR()}}. Differences are:
\itemize{
\item Works with any decomposable graph, not just block graphs
\item Does not support \code{edges_to_add}
}
}
