% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{fmpareto_graph_HR_general}
\alias{fmpareto_graph_HR_general}
\title{Parameter fitting for multivariate Huesler--Reiss Pareto distributions on non-decomposable graphs}
\usage{
fmpareto_graph_HR_general(data, graph, p = NULL, ...)
}
\arguments{
\item{data}{Numeric matrix of size \eqn{n\times d}{n x d}, where \eqn{n} is the
number of observations and \eqn{d} is the dimension.}

\item{graph}{Connected graph object from \code{igraph} package.}

\item{p}{Numeric between 0 and 1 or \code{NULL}. If \code{NULL} (default),
it is assumed that the \code{data} are already on multivariate Pareto scale. Else,
\code{p} is used as the probability in the function \code{\link{data2mpareto}}
to standardize the \code{data}.}

\item{...}{Further arguments passed to \code{\link[=complete_gamma_general]{complete_gamma_general()}} if \code{graph}
is not decomposable}
}
\description{
Works by computing an empirical Gamma matrix and then fitting this to the given \code{graph}.
}
