% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_param.R
\name{fmpareto_graph_HR_clique_average}
\alias{fmpareto_graph_HR_clique_average}
\alias{fmpareto_graph_HR_clique_sequential}
\alias{combine_clique_estimates_by_averaging}
\title{HR Parameter fitting - Helper functions}
\usage{
fmpareto_graph_HR_clique_average(data, graph, method = c("ML", "vario"), ...)

fmpareto_graph_HR_clique_sequential(data, graph, ...)

combine_clique_estimates_by_averaging(cliques, subGammas)
}
\description{
Helper functions called by \code{\link{fmpareto_HR_MLE}}.
}
\keyword{internal}
