% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_plotting.R
\name{formatDegrees}
\alias{formatDegrees}
\title{Helper function to format (coordinate) degrees used as axis labels}
\usage{
formatDegrees(decDeg, direction = "NS", latex = TRUE)
}
\description{
Helper function to format (coordinate) degrees used as axis labels
}
