% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{fpareto_HR}
\alias{fpareto_HR}
\title{Fit parameters of multivariate HR Pareto distribution}
\usage{
fpareto_HR(data, p = NULL, cens = FALSE, init, maxit = 100,
  graph = NULL, method = "BFGS")
}
\arguments{
\item{data}{Numeric matrix \eqn{n\times d}{n x d}. A dataset containing
\eqn{n} observations following a \eqn{d}-variate HR Pareto distribution.}

\item{p}{Numeric between 0 and 1. Threshold probability. If \code{NULL},
it is assumed that \code{data} is already distributed as multivariate Pareto.}

\item{cens}{Boolean. If TRUE, then censored log-likelihood is computed.
By default, \code{cens = FALSE}.}

\item{init}{Numeric vector. Initial parameter values.}

\item{maxit}{Positive integer. The maximum number of iterations in the
optimization.}

\item{graph}{Graph object from \code{igraph} package or
\code{NULL}.
If \code{graph} is not \code{NULL}, then only edge parameters are fit.}

\item{method}{String. A valid optimization method used by the function
\code{\link[stats]{optim}}. By default, \code{method = "BFGS"}.}
}
\value{
List. The list made of:
\itemize{
\item \code{convergence} Boolean. Whether the optimization converged or not.
\item \code{par} Numeric vector. Optimized parameters.
\item \code{Gamma} Numeric matrix \eqn{d \time d}{d x d}. Fitted variogram
matrix.
\item \code{nllik} Numeric. Optimum value of the likelihood function.
\item \code{hessian} Numeric matrix. Estimated Hessian matrix of the
estimated parameters.
}
}
\description{
This function fits the parameters of a multivariate HR Pareto distribution
 using (censored) likelihood estimation.
}
\details{
If \code{graph} is given, it assumes the conditional independence
 structure of this graph and fits only the parameters on the edges,
 but with the full likelihood. This should only be used for small dimensions.
 If you give \code{data}, and not \code{graph}, then the function fits HR
 with the whole matrix. This is computationally heavy and works only for
 small dimensions, e.g., 3 or 4.
}
