% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Gamma2chi}
\alias{Gamma2chi}
\title{Convert \eqn{\Gamma} to \eqn{\chi}}
\usage{
Gamma2chi(Gamma)
}
\arguments{
\item{Gamma}{Numeric matrix \eqn{d\times d}{d x d}.
It represents a variogram matrix.
Alternatively, it can be passed as a a numeric vector with
\eqn{d} elements.}
}
\value{
Numeric. The extremal correlation coefficient.
}
\description{
Transform the \eqn{\Gamma} HR variogram into \code{chi} extremal correlation.
}
