% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations2.R
\name{Gamma2chi_3D}
\alias{Gamma2chi_3D}
\title{Compute theoretical \eChi in 3D}
\usage{
Gamma2chi_3D(Gamma)
}
\arguments{
\item{Gamma}{Numeric \eXTimesY{3}{3} matrix.}
}
\value{
The 3-dimensional \eChi coefficient, i.e.,
the extremal correlation coefficient for the HR distribution. Note that
\eqn{0 \leq \chi \leq 1}.
}
\description{
Computes the theoretical \eChi coefficient in 3 dimensions.
}
\keyword{internal}
