% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Gamma2par}
\alias{Gamma2par}
\title{Extract upper triangular part of \eqn{\Gamma}}
\usage{
Gamma2par(Gamma)
}
\arguments{
\item{Gamma}{Numeric matrix \eqn{d\times d}{d x d}.
It represents a variogram matrix.
Alternatively, it can be passed as a a numeric vector with
\eqn{d} elements.}
}
\value{
Numeric vector with \eqn{d} elements.
The upper triangular part of the given \code{Gamma} matrix.
}
\description{
This function returns a vector containing the upper triangular part
of the matrix \code{Gamma}. If \code{Gamma} is already a vector, it returns
it as it is.
}
