% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.R
\name{generate_random_cactus}
\alias{generate_random_cactus}
\title{Generate a random cactus graph}
\usage{
generate_random_cactus(d, cMin = 2, cMax = 6)
}
\arguments{
\item{d}{Number of vertices in the graph}

\item{cMin}{Minimal size of each block (last block might be smaller)}

\item{cMax}{Maximal size of each block}
}
\description{
Generates a random cactus graph (mostly useful for benchmarking).
}
\seealso{
Other Example generations: 
\code{\link{generate_random_Gamma}()},
\code{\link{generate_random_chordal_graph}()},
\code{\link{generate_random_connected_graph}()},
\code{\link{generate_random_graphical_Gamma}()},
\code{\link{generate_random_integer_Gamma}()},
\code{\link{generate_random_model}()},
\code{\link{generate_random_spd_matrix}()},
\code{\link{generate_random_tree}()}
}
\concept{Example generations}
