% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.R
\name{generate_random_Gamma}
\alias{generate_random_Gamma}
\title{Generate a random Gamma matrix}
\usage{
generate_random_Gamma(d, ...)
}
\arguments{
\item{d}{Size of the matrix}

\item{...}{Further arguments passed to \code{\link[=generate_random_spd_matrix]{generate_random_spd_matrix()}}}
}
\description{
Generates a valid Gamma matrix with a given dimension
}
\seealso{
Other Example generations: 
\code{\link{generate_random_chordal_graph}()},
\code{\link{generate_random_connected_graph}()},
\code{\link{generate_random_graphical_Gamma}()},
\code{\link{generate_random_integer_Gamma}()},
\code{\link{generate_random_model}()},
\code{\link{generate_random_spd_matrix}()},
\code{\link{generate_random_tree}()}
}
\concept{Example generations}
