% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.R
\name{generate_random_graphical_Gamma}
\alias{generate_random_graphical_Gamma}
\title{Generate a random Gamma matrix for a given graph}
\usage{
generate_random_graphical_Gamma(graph, ...)
}
\arguments{
\item{graph}{An \link[igraph:make_graph]{igraph::graph} object}

\item{...}{Furhter arguments passed to \code{\link[=generate_random_spd_matrix]{generate_random_spd_matrix()}}}
}
\description{
Generates a valid Gamma matrix with conditional independence structure
specified by a graph
}
\seealso{
Other Example generations: 
\code{\link{generate_random_chordal_graph}()},
\code{\link{generate_random_connected_graph}()},
\code{\link{generate_random_model}()},
\code{\link{generate_random_spd_matrix}()},
\code{\link{generate_random_tree}()}
}
\concept{Example generations}
