% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.R
\name{generate_random_tree}
\alias{generate_random_tree}
\title{Generate a random tree}
\usage{
generate_random_tree(d)
}
\arguments{
\item{d}{Number of vertices in the graph}
}
\value{
An [\code{igraph::graph}] object
}
\description{
Generates a random tree from a random Pruefer sequence
}
\seealso{
Other example generation functions:
\code{\link{generate_random_Gamma}()},
\code{\link{generate_random_cactus}()},
\code{\link{generate_random_chordal_graph}()},
\code{\link{generate_random_connected_graph}()},
\code{\link{generate_random_graphical_Gamma}()},
\code{\link{generate_random_integer_Gamma}()},
\code{\link{generate_random_model}()},
\code{\link{generate_random_spd_matrix}()}
}
\concept{exampleGenerations}
