% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_alert_function}
\alias{get_alert_function}
\title{Get alert function}
\usage{
get_alert_function(overwrite = NULL)
}
\arguments{
\item{overwrite}{\code{NULL} or \code{TRUE} to read the option value,
\code{FALSE} to return a dummy function,
or a function that takes an arbitrary number of strings as arguments (e.g. \code{stop()}).}
}
\value{
A function that takes an arbitrary number of strings as arguments.
}
\description{
Get a function that can be used to alert the user of invalid inputs.
Returns the value implied by the \code{overwrite} argument,
or the option \code{"graphicalExtremes.default.alert"},
falling back to \code{warning()} if neither is specified.
}
\seealso{
\code{\link{graphicalExtremes-package}}

Other getOptions: 
\code{\link{get_mc_cores}()},
\code{\link{get_small_tol}()}
}
\concept{getOptions}
