% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions_2.R
\name{get_cliques_and_separators}
\alias{get_cliques_and_separators}
\title{Get Cliques and Separators of a graph}
\usage{
get_cliques_and_separators(graph)
}
\arguments{
\item{graph}{An \link[igraph:make_graph]{igraph::graph} object}
}
\value{
A list of vertex sets that represent the cliques and (recursive)
separators of \code{graph}
}
\description{
Finds all cliques, separators, and (recursively) separators of separators
in a graph.
}
