% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{get_cliques_and_separators_OLD}
\alias{get_cliques_and_separators_OLD}
\title{Get Cliques and Separators of a graph}
\usage{
get_cliques_and_separators_OLD(graph)
}
\arguments{
\item{graph}{An [\code{igraph::graph}] object}
}
\value{
A list of vertex sets that represent the cliques and (recursive)
separators of \code{graph}
}
\description{
Finds all cliques, separators, and (recursively) separators of separators
in a graph.
}
\keyword{internal}
