% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{get_critical_ev_Sigma_Theta}
\alias{get_critical_ev_Sigma_Theta}
\title{Assumest one 0-eigenvalue and returns the smallest other eigenvalue of a matrix}
\usage{
get_critical_ev_Sigma_Theta(M)
}
\description{
Assumest one 0-eigenvalue and returns the smallest other eigenvalue of a matrix
}
