% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{get_second_smallest_ev}
\alias{get_second_smallest_ev}
\title{Returns the second smallest eigenvalue of a matrix}
\usage{
get_second_smallest_ev(M)
}
\description{
Returns the second smallest eigenvalue of a matrix
}
