% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_tol}
\alias{get_tol}
\title{Tolerance to be used instead of machine precision}
\usage{
get_tol(overwrite = NULL)
}
\arguments{
\item{overwrite}{Use this value if specified}
}
\value{
A non-negative numerical scalar
}
\description{
Helper function that returns the tolerance to be used e.g. when checking
matrices for symmetry and definiteness.
Can be set using \code{setOption('graphicalExtremes.tol', ...)}.
}
\details{
In general, this value is used only as a "permissive" tolerance, in the sense
that if a value has to be positive, it is compared to actual zero, but if
it has to be zero, its absolute value is compared to this tolerance.
}
