% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_danube.R
\name{getDanubeFlowGraph}
\alias{getDanubeFlowGraph}
\title{Get Danube flow graph}
\usage{
getDanubeFlowGraph(stationIndices = NULL, directed = FALSE)
}
\arguments{
\item{stationIndices}{Logical or numerical vector. Indicating which stations to include.}

\item{directed}{Logical. Whether the graph should be directed (in the direction of flow).}
}
\value{
An \code{\link[igraph:make_graph]{igraph::graph}} object.
}
\description{
Returns an \code{\link[igraph:make_graph]{igraph::graph}} object representing the flow graph of the \code{\link{danube}} dataset.
}
