% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flight_plotting.R
\name{getFlightDelayData}
\alias{getFlightDelayData}
\title{Get delay data}
\usage{
getFlightDelayData(
  what = c("delays", "IATAs", "dates"),
  IATAfilter = c("all", "tcCluster", "tcAll"),
  dateFilter = c("all", "tcTrain", "tcTest")[1],
  delayFilter = c("totals", "arrivals", "departures")[1]
)
}
\description{
Get filtered delay data.
}
