% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getPackageData}
\alias{getPackageData}
\title{Get package data}
\usage{
getPackageData(filename, isRDS = TRUE)
}
\arguments{
\item{filename}{Base name of the file. E.g. \code{"FILENAME"} for a file \code{inst/extdata/FILENAME}.}

\item{isRDS}{Whether the file is an \code{.RDS} file.}
}
\value{
If \code{isRDS=TRUE} an R object.
If \code{isRDS=FALSE} an environment, containing the R objects from the file.
}
\description{
Get private data sets from \code{inst/extdata}.
}
\keyword{internal}
