% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphUtils.R
\name{getSubMatrixForSubgraph}
\alias{getSubMatrixForSubgraph}
\title{Get the submatrix corresponding to a subgraph}
\usage{
getSubMatrixForSubgraph(fullMatrix, subgraph, graph = NULL)
}
\description{
Both the graph and subgraph need to have persistent IDs
If graph==NULL it is assumed to have pIDs 1, 2, ...
}
