% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_helpers.R
\name{graphs_equal}
\alias{graphs_equal}
\title{Graph equality}
\usage{
graphs_equal(g1, g2)
}
\arguments{
\item{g1}{\code{igraph::graph}}

\item{g2}{\code{igraph::graph}}
}
\value{
\code{logical} indicating if the graphs are equal
}
\description{
Produce true if two graphs have same vertices and edges (ordered)
}
\keyword{internal}
