% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{is_block_graph}
\alias{is_block_graph}
\title{Check if a graph is a block graph}
\usage{
is_block_graph(graph)
}
\arguments{
\item{graph}{An \link[igraph:make_graph]{igraph::graph} object}
}
\value{
A \code{boolean} indicating if the graph is a glock graph
}
\description{
Check if a graph is a block graph
}
\seealso{
Other Input checks: 
\code{\link{check_Gamma_and_graph}()},
\code{\link{check_graph}()}
}
\concept{Input checks}
