% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{is_cnd}
\alias{is_cnd}
\title{Check if a square, symmetric matrix is conditionally negative definite}
\usage{
is_cnd(M)
}
\description{
Check if a square, symmetric matrix is conditionally negative definite
}
