% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{is_symmetric_matrix}
\alias{is_symmetric_matrix}
\title{Check if an object is a square, symmetric matrix}
\usage{
is_symmetric_matrix(M, tol = get_small_tol())
}
\description{
Check if an object is a square, symmetric matrix
}
