% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions_general.R
\name{iterateIndList}
\alias{iterateIndList}
\title{Workhorse of \code{complete_Gamma_general()}}
\usage{
iterateIndList(Gamma, sepList, nonEdgeIndices, N, tol, check_tol)
}
\description{
Workhorse of \code{complete_Gamma_general()}
}
