% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helpers.R
\name{logdV_HR}
\alias{logdV_HR}
\title{Compute the exponent measure density of HR distribution}
\usage{
logdV_HR(x, Gamma = NULL, Theta = NULL)
}
\arguments{
\item{x}{Numeric matrix \eqn{n\times d}{n x d} or vector with \eqn{d} elements.}
}
\value{
Numeric. The censored exponent measure of the HR distribution.
}
\description{
Computes the exponent measure density of HR distribution.
}
\keyword{internal}
