% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{loglik_HR}
\alias{loglik_HR}
\title{Compute Huesler--Reiss log-likelihood, AIC, and BIC}
\usage{
loglik_HR(data, p = NULL, graph, Gamma, cens = FALSE)
}
\arguments{
\item{data}{Numeric matrix \eqn{n\times d}{n x d}. It contains
observations following a multivariate HR Pareto distribution.}

\item{p}{Numeric between 0 and 1 or \code{NULL}. If \code{NULL} (default),
it is assumed that the \code{data} are already on multivariate Pareto scale.
Else, \code{p} is used as the probability in the function \code{\link{data2mpareto}}
to standardize the \code{data}.}

\item{graph}{An \link[igraph:make_graph]{igraph::graph} object. The \code{graph} must be undirected and
connected}

\item{Gamma}{Numeric matrix \eqn{n\times d}{n x d}.
It represents a variogram matrix \eqn{\Gamma}.}

\item{cens}{Boolean. If true, then censored log-likelihood is computed.
By default, \code{cens = FALSE}.}
}
\value{
Numeric vector \code{c("loglik"=..., "aic"=..., "bic"=...)} with the evaluated
log-likelihood, AIC, and BIC values.
}
\description{
Computes (censored) Huesler--Reiss log-likelihood, AIC, and BIC values.
}
