% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_graph_list}
\alias{make_graph_list}
\title{Create graph list for non-decomposable completion}
\usage{
make_graph_list(graph)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.}
}
\value{
List of decomposable graphs
}
\description{
Creates a list of decomposable graphs, each consisting of two cliques,
such that the intersection of their edge sets
is identical to the edgeset of the input graph.
}
