% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_numeric_indices}
\alias{make_numeric_indices}
\title{Convert indices to numerical indices}
\usage{
make_numeric_indices(ind, n = NULL, unique = TRUE, sort = TRUE)
}
\arguments{
\item{ind}{The numerical or logical index vector}

\item{n}{Max numerical index (used if \code{ind} is logical and might be recycled)}

\item{unique}{Whether to keep every (numerical) index at most once}

\item{sort}{Whether to sort the numerical indices}
}
\value{
A numerical index vector
}
\description{
Converts (possibly) logical indices to numerical ones.
Also ensures unique indices and sorts them if specified.
}
\keyword{internal}
