% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{make_sep_list}
\alias{make_sep_list}
\title{Create a list of separators}
\usage{
make_sep_list(graph, details = TRUE)
}
\arguments{
\item{graph}{A graph}
}
\value{
A list of numeric vectors
}
\description{
Creates a list of separator set, such that every pair of non-adjacent
vertices in \code{graph} is completely disconnected by the removal of
(at least) one of the separator sets from the graph.
}
