% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{max_without_warning}
\alias{max_without_warning}
\title{Compute the maximum without warning about \code{max(c()) = -Inf}}
\usage{
max_without_warning(..., na.rm = FALSE)
}
\description{
Compute the maximum without warning about \code{max(c()) = -Inf}
}
