% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{min_without_warning}
\alias{min_without_warning}
\title{Compute the minimum without warning about \code{min(c()) = Inf}}
\usage{
min_without_warning(..., na.rm = FALSE)
}
\description{
Compute the minimum without warning about \code{min(c()) = Inf}
}
