% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{mst_HR}
\alias{mst_HR}
\title{Estimate the HR minimum spanning tree}
\usage{
mst_HR(data, p = NULL, cens = FALSE)
}
\arguments{
\item{data}{Numeric matrix \eqn{n\times d}{n x d}. It contains
observations following a multivariate HR Pareto distribution.}

\item{p}{Numeric between 0 and 1. Threshold probability. If \code{NULL},
it is assumed that \code{data} is already distributed as multivariate Pareto.}

\item{cens}{Boolean. If TRUE, then censored log-likelihood is computed.
By default, \code{cens = FALSE}.}
}
\value{
Graph object from \code{igraph} package. A tree.
}
\description{
Estimates the minimum spanning tree as the HR tree that maximizes
the (censored) log-likelihood.
}
