% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{order_cliques}
\alias{order_cliques}
\title{Order Cliques}
\usage{
order_cliques(cliques)
}
\description{
Orders the cliques in a connected decomposable graph so that they fulfill the running intersection property.
}
