% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{par2Gamma}
\alias{par2Gamma}
\title{Create \eqn{\Gamma} from vector}
\usage{
par2Gamma(par)
}
\arguments{
\item{par}{Numeric vector with \eqn{d} elements.
Upper triangular part of a Gamma matrix.}
}
\value{
Numeric matrix \eqn{d \times d}{d x d}. Full Gamma matrix.
}
\description{
This function takes the parameters in the vector \code{par}
(upper triangular Gamma matrix) and returns full Gamma.
}
\keyword{internal}
