% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations2.R
\name{Theta2Theta}
\alias{Theta2Theta}
\alias{Sigma2Sigma}
\alias{Gamma2Sigma}
\alias{Gamma2Theta}
\alias{Sigma2Gamma}
\alias{Sigma2Theta}
\alias{Theta2Sigma}
\title{Convert between representations of \eTheta, \eThetaK / \eSigma, \eSigmaK}
\usage{
Theta2Theta(
  Theta,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Sigma2Sigma(
  Sigma,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Gamma2Sigma(Gamma, k = NULL, full = FALSE, check = TRUE)

Gamma2Theta(Gamma, k = NULL, full = FALSE, check = TRUE)

Sigma2Gamma(Sigma, k = NULL, full = FALSE, check = TRUE)

Sigma2Theta(
  Sigma,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Theta2Sigma(
  Theta,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)
}
\arguments{
\item{Theta}{Numeric \dxd or \d1xd1 precision matrix.}

\item{k1}{\code{NULL} if the input matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{k2}{\code{NULL} if the output matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{full1}{Logical. If \code{TRUE} and \code{!is.null(k1)},
the input is a \dxd matrix with the kth row filled with zeros.}

\item{full2}{Logical. If \code{TRUE} and \code{!is.null(k2)},
the output is a \dxd matrix with the kth row filled with zeros.}

\item{check}{Whether to check the inputs and ensure numerical symmetry of outputs.}

\item{Sigma}{Numeric \dxd or \d1xd1 covariance matrix.}

\item{Gamma}{Numeric \dxd variogram matrix.}

\item{k}{\code{NULL} if the input/output matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{full}{Logical. If \code{TRUE} and \code{!is.null(k)},
the input/output matrix is a \dxd matrix with the kth row filled with zeros.}
}
\value{
The desired parameter matrix corresponding to the specified inputs.
}
\description{
Converts between matrices \eTheta and \eThetaK,
possibly for different values of k,
and possibly filling the kth row/colum with zeros.
Same for \eSigma, \eSigmaK.
}
\details{
If \code{k}, \code{k1}, or \code{k2} is \code{NULL}, the corresponding \code{full_} argument is ignored.
}
