% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{partialMatrixToGraph}
\alias{partialMatrixToGraph}
\title{Transformation of a partial matrix to a graph}
\usage{
partialMatrixToGraph(Matrix)
}
\arguments{
\item{Matrix}{A matrix with NA entries}
}
\value{
An \code{igraph::graph} object
}
\description{
Creates a graph that has edges in entries corresponding to non-NA entries
in Gamma.
}
