% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rmpareto_tree}
\alias{rmpareto_tree}
\title{Sampling of a multivariate Pareto distribution on a tree}
\usage{
rmpareto_tree(n, model = c("HR", "logistic", "dirichlet")[1], tree, par)
}
\arguments{
\item{n}{Number of simulations.}

\item{model}{The parametric model type; one of:
\itemize{
\item \code{HR} (default),
\item \code{logistic},
\item \code{dirichlet}.
}}

\item{tree}{Graph object from \code{igraph} package.
This object must be tree, i.e., an
undirected graph that is connected and has no cycles.}

\item{par}{Respective parameter for the given \code{model}, that is,
\itemize{
\item \eqn{\Gamma}, numeric \eqn{d \times d}{d x d} variogram matrix,
where only the entries corresponding to the edges of the \code{tree} are used,
if \code{model = HR}. Alternatively, can be a vector of
length \eqn{d - 1} containing the entries of the variogram corresponding
to the edges of the given \code{tree}.
\item \eqn{\theta \in (0, 1)}{0 < \theta < 1}, vector of length \eqn{d - 1}
containing the logistic parameters corresponding
to the edges of the given \code{tree}, if \code{model = logistic}.
\item a matrix of size \eqn{(d - 1) \times 2}{(d - 1) x 2}, where the rows
contain the parameters vectors \eqn{\alpha} of size 2 with positve entries
for each of the edges in \code{tree}, if \code{model = dirichlet}.
}}
}
\value{
Numeric matrix of size \eqn{n \times d}{n x d}.
The simulated multivariate Pareto data.

## !!! add examples (define params and call function)
}
\description{
Simulates exact samples of a multivariate Pareto distribution that
is an extremal graphical model on a tree as defined in CITE eng2018a.
}
