% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rmstable}
\alias{rmstable}
\title{Simulate samples of max-stable process}
\usage{
rmstable(n, model = c("HR", "logistic", "neglogistic", "dirichlet")[1],
  d, par)
}
\arguments{
\item{n}{Positive integer. Number of simulations.}

\item{model}{String. The parametric model type. Is one of:
\itemize{
\item \code{HR} (default),
\item \code{logistic},
\item \code{neglogistic},
\item \code{dirichlet}.
}}

\item{d}{Positive integer. Dimension of the multivariate Pareto
distribution.}

\item{par}{Is the respective parameter for the given \code{model}.
Is one of:
\itemize{
\item \eqn{\Gamma}, numeric matrix representing a \eqn{d \times d}{d x d}
variogram, if \code{model = HR}.
\item \eqn{\theta \in (0, 1)}{0 < \theta < 1}, if \code{model = logistic}.
\item \eqn{\theta > 0}, if \code{model = neglogistic}.
\item \eqn{\alpha > 0}, numeric vector of size \code{d},
if \code{model = dirichlet}.
}}
}
\value{
List. The list is made of:
\itemize{
\item \code{res} Numeric matrix of size \eqn{n \times d}{n x d}.
The simulated multivariate Pareto data.
\item \code{counter} Positive integer. The number of times needed to sweep
over the \code{d} variables to simulate \code{n} multivariate
observations.
}
## !!! add examples (define params and call function)
}
\description{
Simulates exact samples of max-stable process
}
