% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rmstable_tree}
\alias{rmstable_tree}
\title{Simulate samples of max-stable process from a tree}
\usage{
rmstable_tree(n, model = c("HR", "logistic", "dirichlet")[1], tree, par)
}
\arguments{
\item{n}{Positive integer. Number of simulations.}

\item{model}{String. The parametric model type. Is one of:
\itemize{
\item \code{HR} (default),
\item \code{logistic},
\item \code{dirichlet}.
}}

\item{tree}{igraph object. An igraph object representing a tree, i.e., an
undirected graph that is connected and has no cycles.}

\item{par}{Is the respective parameter for the given \code{model}.
Is one of:
\itemize{
\item \eqn{\Gamma}, numeric matrix representing a \eqn{d \times d}{d x d}
variogram, if \code{model = HR}. Alternatively, can be a vector of
length \eqn{d - 1}, containing the entries of the variogram corresponding
to the edges of the given \code{tree}.
\item \eqn{\theta \in (0, 1)}{0 < \theta < 1}, if \code{model = logistic}
\item a matrix of size \eqn{(d - 1) \times 2}{(d - 1) x 2} containing
the relative \eqn{\alpha > 0} coefficients, if \code{model = dirichlet}.
}}
}
\value{
List. The list is made of:
\itemize{
\item \code{res} Numeric matrix of size \eqn{n \times d}{n x d}.
The simulated multivariate Pareto data.
\item \code{counter} Positive integer. The number of times needed to sweep
over the \code{d} variables to simulate \code{n} multivariate
observations.
}
## !!! add examples (define params and call function)
}
\description{
Simulates a tree graphical model, following a max-stable process.
}
