% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{select_edges}
\alias{select_edges}
\title{Select edges to add to a graph}
\usage{
select_edges(graph)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.
The \code{graph} must be an undirected block graph, i.e., a decomposable, connected
graph with singleton separator sets.}
}
\value{
Numeric vector.
}
\description{
This function selects all possible edges that can be added to graph
while still remaining in the class of block graphs.
}
