% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{selectEdges}
\alias{selectEdges}
\title{Select edges to add to a graph}
\usage{
selectEdges(graph)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.
An undirected block graph, i.e., a decomposable, connected
graph where the minimal separators of the cliques have size at most one.}
}
\value{
Numeric vector.
}
\description{
This function selects all possible edges that can be added to graph
while still remaining in the class of block graphs.
}
