% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{set_graph_parameters}
\alias{set_graph_parameters}
\title{Set graphical parameters}
\usage{
set_graph_parameters(graph)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.}
}
\value{
Graph object from \code{igraph} package.
}
\description{
Set graphical parameters to \code{graph} which is an object from the
\code{igraph} package.
}
\keyword{internal}
