% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simu_px_dirichlet}
\alias{simu_px_dirichlet}
\title{Simulate Dirichlet extremal functions}
\usage{
simu_px_dirichlet(n, d, idx, alpha)
}
\arguments{
\item{n}{Positive integer. Number of simulations.}

\item{d}{Positive integer. Dimension of the multivariate Pareto
distribution.}

\item{idx}{Integer or numeric vector with \code{n} elements. Inde(x|ces) from
1 to \code{d}, that determine which extremal function to simulate.}

\item{alpha}{Numeric vector of size \code{d}.}
}
\value{
Numeric matrix \eqn{n\times d}{n x d}. Simulated data.
}
\description{
Simulates Dirichlet extremal functions
}
