% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simu_px_HR}
\alias{simu_px_HR}
\title{Simulates HR extremal functions}
\usage{
simu_px_HR(no.simu = 1, idx, trend, chol.mat)
}
\arguments{
\item{no.simu}{Integer. Number of simulations, by default equal to 1}

\item{idx}{Integer.}
}
\description{
Simulate the Huessler-Reiss extremal functions
}
