% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simu_px_logistic}
\alias{simu_px_logistic}
\title{Simulate logistic extremal functions}
\usage{
simu_px_logistic(n, d, idx, theta)
}
\arguments{
\item{n}{Number of simulations.}

\item{d}{Dimension of the multivariate Pareto
distribution.}

\item{idx}{Integer or numeric vector with \code{n} elements. Inde(x|ces) from
1 to \code{d}, that determine which extremal function to simulate.}

\item{theta}{Numeric --- assume \eqn{0 < \theta < 1}.}
}
\value{
Numeric \nxd matrix. Simulated data.
}
\description{
Simulates logistic extremal functions
}
\keyword{internal}
