% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simu_px_neglogistic}
\alias{simu_px_neglogistic}
\title{Simulate negative logistic extremal functions}
\usage{
simu_px_neglogistic(n, d, idx, theta)
}
\arguments{
\item{n}{Positive integer. Number of simulations.}

\item{d}{Positive integer. Dimension of the multivariate Pareto
distribution.}

\item{idx}{Integer or numeric vector with \code{n} elements. Inde(x|ces) from
1 to \code{d}, that determine which extremal function to simulate.}

\item{theta}{Numeric --- assume \eqn{\theta > 0}.}
}
\value{
Numeric matrix \eqn{n\times d}{n x d}. Simulated data.
}
\description{
Simulates negative logistic extremal functions
}
