% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{sort_cliques_and_separators}
\alias{sort_cliques_and_separators}
\title{Sort a set of cliques and separators into layers, such that the separator
of two cliques/separators is contained in a previous layer}
\usage{
sort_cliques_and_separators(cliques)
}
\description{
Sort a set of cliques and separators into layers, such that the separator
of two cliques/separators is contained in a previous layer
}
