% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{split_graph}
\alias{split_graph}
\title{Split graph into invariant subgraphs}
\usage{
split_graph(g)
}
\arguments{
\item{g}{\code{\link[igraph:make_graph]{igraph::graph}} object}
}
\value{
List of invariant subgraphs
}
\description{
Split graph into invariant subgraphs
}
\keyword{internal}
