% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs_functions.R
\name{split_graph_at_sep}
\alias{split_graph_at_sep}
\title{Split a graph at a single separator}
\usage{
split_graph_at_sep(
  g,
  sepIds = NULL,
  sepPids = NULL,
  returnGraphs = FALSE,
  includeSep = TRUE
)
}
\description{
Split a graph at a single separator
}
