% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Theta2Gamma}
\alias{Theta2Gamma}
\title{Transformation of \eqn{\Gamma} matrix to \eqn{\Theta} matrix}
\usage{
Theta2Gamma(Theta)
}
\description{
Transformation of \eqn{\Gamma} matrix to \eqn{\Theta} matrix
}
