% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations2.R
\name{Theta2Theta}
\alias{Theta2Theta}
\alias{Sigma2Sigma}
\title{Convert between different representations of \eTheta / \eSigma}
\usage{
Theta2Theta(
  Theta,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)

Sigma2Sigma(
  Sigma,
  k1 = NULL,
  k2 = NULL,
  full1 = FALSE,
  full2 = FALSE,
  check = TRUE
)
}
\arguments{
\item{Theta}{Precision matrix \eTheta or \eThetaK.}

\item{k1}{\code{NULL} if the input matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{k2}{\code{NULL} if the output matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{full1}{Logical. If \code{TRUE} and \code{!is.null(k1)}, the input is a \dxd matrix with the kth row filled with zeros.}

\item{full2}{Logical. If \code{TRUE} and \code{!is.null(k2)}, the output is a \dxd matrix with the kth row filled with zeros.}

\item{check}{Whether to check the inputs and ensure numerical symmetry of outputs.}

\item{Sigma}{Covariance matrix \eSigma or \eSigmaK.}
}
\description{
Convert between different representations of \eTheta / \eSigma
}
