% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{unif}
\alias{unif}
\title{Uniform margin}
\usage{
unif(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{Logical. If TRUE, missing values are removed. If FALSE, missing values are kept as such.}
}
\value{
Numeric vector with entries rescaled to uniform margins
}
\description{
Rescale the vector \code{x} empirically to uniform margin.
}
\keyword{internal}
