% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helpers.R
\name{V_HR}
\alias{V_HR}
\title{Compute exponent measure}
\usage{
V_HR(x, Gamma = NULL, Theta = NULL)
}
\arguments{
\item{x}{Numeric vector with \eqn{d} positive elements
where the exponent measure is to be evaluated.}

\item{par}{\eqn{d \times d}{d x d} variogram matrix, or numeric vector with
\eqn{\frac{d(d - 1)}{2}}{d x (d - 1) / 2} elements,
representing the upper triangular portion of a
variogram matrix \eqn{\Gamma}.}
}
\value{
Numeric. The exponent measure of the HR distribution.
}
\description{
Computes the exponent measure of HR distribution.
}
\keyword{internal}
