% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_helpers.R
\name{censor}
\alias{censor}
\title{Censor dataset}
\usage{
censor(x, p)
}
\arguments{
\item{x}{Numeric matrix \eqn{n \times d}{n x d}.}

\item{p}{Numeric vector with \eqn{d} elements.}
}
\value{
Numeric matrix \eqn{n \times d}{n x d}.
}
\description{
Censors each row of matrix \code{x} with vector \code{p}.
}
\keyword{internal}
