% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_checks.R
\name{checkGamma}
\alias{checkGamma}
\alias{checkSigmaTheta}
\alias{checkTheta}
\alias{checkSigma}
\alias{checkMatrix}
\title{Parameter matrix checks}
\usage{
checkGamma(Gamma, alert = NULL, tol = get_small_tol(), returnBoolean = FALSE)

checkSigmaTheta(
  M,
  k,
  full,
  matrixName = "Sigma",
  tol = get_small_tol(),
  alert = NULL,
  returnBoolean = FALSE
)

checkTheta(
  Theta,
  k = NULL,
  full = FALSE,
  tol = get_small_tol(),
  alert = NULL,
  returnBoolean = FALSE
)

checkSigma(
  Sigma,
  k = NULL,
  full = FALSE,
  tol = get_small_tol(),
  alert = NULL,
  returnBoolean = FALSE
)

checkMatrix(
  M,
  name = c("Gamma", "Sigma", "Theta")[1],
  k = NULL,
  full = FALSE,
  tol = get_small_tol(),
  alert = NULL,
  returnBoolean = FALSE
)
}
\arguments{
\item{Gamma}{Numeric \dxd variogram matrix.}

\item{tol}{Numeric scalar. Entries in the precision matrix with absolute value
smaller than this are considered to be zero.}

\item{M}{Numeric matrix, \eGamma, \eSigma, or \eTheta.}

\item{k}{\code{NULL} if the input/output matrix is \eSigma/\eTheta.
Else, an integer between 1 and d indicating the value of k in \eSigmaK, \eThetaK.}

\item{full}{Logical. If \code{TRUE} and \code{!is.null(k)},
the input/output matrix is a \dxd matrix with the kth row filled with zeros.}

\item{Theta}{Numeric \dxd or \d1xd1 precision matrix.}

\item{Sigma}{Numeric \dxd or \d1xd1 covariance matrix.}
}
\value{
The input matrix, passed through \code{\link{ensure_matrix_symmetry_and_truncate_zeros}},
if it is ok, or throws an error, otherwise.
}
\description{
Checks wheter the matrix given is a valid Huesler-Reiss parameter matrix
in the role of \eGamma, \eTheta, or \eSigma, respectively.
}
