% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{chi2Gamma}
\alias{chi2Gamma}
\title{Convert \eqn{\chi} to \eqn{\Gamma}}
\usage{
chi2Gamma(chi)
}
\arguments{
\item{chi}{Numeric or matrix, with entries
between 0 and 1.}
}
\value{
Numeric
}
\description{
Transform the \code{chi} extremal correlation into HR parameter.
}
