% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations.R
\name{chi2Gamma}
\alias{chi2Gamma}
\title{Transformation of extremal correlation \eChi to the Huesler--Reiss variogram \eGamma}
\usage{
chi2Gamma(chi)
}
\arguments{
\item{chi}{Numeric or matrix, with entries
between 0 and 1.}
}
\value{
Numeric or matrix. The \eGamma parameters in the Huesler--Reiss
distribution.
}
\description{
Transforms the extremal correlation \eChi into the \code{Gamma} matrix
from the definition of a Huesler--Reiss distribution.
}
\details{
The formula for transformation from \code{chi} to \eGamma that is applied element-wise is
\deqn{\Gamma = (2 \Phi^{-1}(1 - 0.5 \chi))^2,}
where \eqn{\Phi^{-1}} is the inverse of the standard normal distribution function.
This is the inverse of \code{\link[=Gamma2chi]{Gamma2chi()}}.
}
