% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions.R
\name{complete_Gamma}
\alias{complete_Gamma}
\title{Completion of Gamma matrices}
\usage{
complete_Gamma(Gamma, graph = NULL, allowed_graph_type = "general", ...)
}
\arguments{
\item{Gamma}{Numeric \eqn{d \times d}{d x d} variogram matrix.}

\item{...}{Further arguments passed to \code{\link[=complete_gamma_general]{complete_gamma_general()}} if \code{graph}
is not decomposable}
}
\value{
Completed \eqn{d \times d}{d x d} \code{Gamma} matrix.
}
\description{
Given a \code{graph} and \code{Gamma} matrix specified (at least) on the
edges of \code{graph}, returns the full \code{Gamma} matrix implied
by the conditional independencies.
}
\details{

}
\examples{
## Complete a 4-dimensional HR distribution

my_graph <- igraph::graph_from_adjacency_matrix(rbind(
  c(0, 1, 0, 0),
  c(1, 0, 1, 1),
  c(0, 1, 0, 1),
  c(0, 1, 1, 0)
),
mode = "undirected"
)

Gamma <- rbind(
  c(0, .5, NA, NA),
  c(.5, 0, 1, 1.5),
  c(NA, 1, 0, .8),
  c(NA, 1.5, .8, 0)
)

complete_Gamma(Gamma, my_graph)

## Alternative

Gamma_vec <- c(.5, 1, 1.5, .8)
complete_Gamma(Gamma_vec, my_graph)
}
\references{
\insertAllCited{}
}
