% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_completions_main.R
\name{complete_Gamma_one_step}
\alias{complete_Gamma_one_step}
\title{Completion of two-clique decomposable Gamma matrices}
\usage{
complete_Gamma_one_step(Gamma, nA, nC, nB)
}
\arguments{
\item{Gamma}{A variogram matrix that is specified on the edges of \code{graph}
and the diagonals. All other entries are ignored.}

\item{graph}{A decomposable \link[igraph:make_graph]{igraph::graph} object with two cliques and
non-empty separator set}
}
\value{
A complete variogram matrix that agrees with \code{Gamma} on the entries
corresponding to edges in \code{graph} and the diagonals.
The corresponding \eTheta matrix produced by \code{\link[=Gamma2Theta]{Gamma2Theta()}} has zeros
in the remaining entries.
}
\description{
Given a decomposable \code{graph} consisting of two cliques and incomplete
variogram matrix \code{Gamma},
returns the full \code{Gamma} matrix implied by the conditional independencies.
}
\seealso{
Other Matrix completions: 
\code{\link{DEPRECATED_complete_Gamma_general}()},
\code{\link{complete_Gamma_decomposable}()},
\code{\link{complete_Gamma_general_demo}()},
\code{\link{complete_Gamma}()}
}
\concept{Matrix completions}
