% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danube}
\alias{danube}
\title{Upper Danube basin dataset}
\format{
An object of class \code{list} of length 4.
}
\source{
Bavarian Environmental Agency \url{http://www.gkd.bayern.de}.
}
\usage{
danube
}
\description{
A dataset containing river discharge data for tributaries of Danube.
}
\keyword{datasets}
