% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danube}
\alias{danube}
\title{Upper Danube basin dataset.}
\format{
A list containing:
\describe{
\item{DataEvents}{a dataframe with the declustered timeseries for the 31
stations.}
...
}
}
\source{
???
}
\usage{
danube
}
\description{
A dataset containing river discharge data for tributaries of Danube in the ...
}
\keyword{datasets}
