% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{danube}
\alias{danube}
\title{Upper Danube basin dataset}
\format{
A named \code{list} with four entries
\describe{
\item{\code{data}}{A numeric matrix, containing discharge data for each gauging station}
\item{\code{info}}{A data frame, containing information about each gauging station}
\item{\code{flow_edges}}{
A two-column numeric matrix. Each row contains the indices (in \code{info})
of a pair of gauging stations that are directly connected by a river.
}
\item{\code{coords_to_plot}}{
A two-column numeric matrix, containing X-Y-coordinates which can be used
to arrange the gauging stations when plotting a flow graph.
}
}
}
\source{
Bavarian Environmental Agency \url{http://www.gkd.bayern.de}.
}
\usage{
danube
}
\description{
A dataset containing river discharge data for tributaries of the Danube.
}
\details{
To obtain the matrix \code{data}, daily discharge data from the summer months of
1960 to 2010 was declustered, yielding between seven and ten observations per year.
Each row corresponds to one observation from this declustered time series,
the non-unique rownames indicate which year an observation is from.
Each column corresponds to one of the gauging stations,
with column indices in \code{data} corresponding to row indices in \code{info}.
See Asadi et al. (2015) for details on the preprocessing and declustering.

\code{info} is a data frame containing the following information for
each of the gauging stations or its corresponding catchment area.
\strong{TODO}: Verify/Delete these.
\describe{
\item{\code{RivNames}}{Name of the river at the gauging station}
\item{\code{Lat}, \code{Long}}{Coordinates of the gauging station}
\item{\code{AveVol}}{\strong{TODO}: Delete?: Some kind of average flow volume, I suppose...}
\item{\code{Lat_Center}, \code{Long_Center}}{Coordinates of the center of the catchment corresponding to the gauging station}
\item{\code{Alt}}{Mean altitude of the catchment}
\item{\code{Area}}{Area of the catchment corresponding to the gauging station}
\item{\code{Chos}}{\strong{TODO}: Delete?: Indices in some larger list of stations.}
\item{\code{Density}}{\strong{TODO}: Delete?: Suggestion by Christina:
"Discharge density means the volume of effluent discharged per unit of time,
per unit area of land available to assimilate the discharge"
}
\item{\code{Slope}}{Mean slope of the catchment}
}
}
\references{
Peiman Asadi, Anthony C. Davison, Sebastian Engelke. "Extremes on river networks."
The Annals of Applied Statistics, 9(4) 2023-2050 December 2015. \url{https://doi.org/10.1214/15-AOAS863}.
}
\seealso{
\code{\link{flights}}
}
\keyword{datasets}
