% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{eglasso}
\alias{eglasso}
\title{DEPRECATED: Fitting extremal graphical lasso}
\usage{
eglasso(
  Gamma,
  rholist = c(0.1, 0.15, 0.19, 0.205),
  reg_method = c("mb", "glasso"),
  eps = 0.5,
  complete_Gamma = FALSE
)
}
\arguments{
\item{Gamma}{Numeric matrix \eqn{n\times d}{n x d}.
It represents a variogram matrix \eqn{\Gamma}.}

\item{rholist}{Numeric vector of non-negative regularization parameters
for the lasso. For details see \link[glasso:glassopath]{glasso::glassopath}.}

\item{reg_method}{One of \code{"mb"} and \code{"glasso"}.
Default is \code{reg_method = "mb"}.}

\item{eps}{Regularization parameter for the covariance matrix.
Default is \code{eps = 0.5}.}

\item{complete_Gamma}{Whether you want to complete Gamma matrix.
Default is \code{complete_Gamma = FALSE}.}
}
\value{
List made of:
\describe{
\item{\code{graph}}{A list of \link[igraph:make_graph]{igraph::graph} objects representing the
fitted graphs for each \code{rho} in \code{rholist}.}
\item{\code{Gamma}}{A list of numeric \eqn{d\times d}{d x d} estimated
variogram matrices \eqn{\Gamma} corresponding to the fitted graphs,
for each \code{rho} in \code{rholist}.}
\item{\code{rholist}}{The list of penalty coefficients.}
}
}
\description{
Fits an extremal minimum spanning tree
}
