% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eglatent.R
\name{eglatent}
\alias{eglatent}
\title{Learning extremal graph structure with latent variables}
\usage{
eglatent(
  Gamma,
  lam1_list = c(0.1, 0.15, 0.19, 0.205),
  lam2_list = c(0.1, 0.15, 0.19, 0.205),
  refit = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Gamma}{Numeric matrix, ??}

\item{lam1_list}{Numeric vector, ??}

\item{lam2_list}{Numeric vector, ??}

\item{refit}{Logical scalar, ??}

\item{verbose}{Logical scalar, indicating whether to print progress updates.}
}
\value{
List made of:
\item{\code{G_obs}}{
List of numeric matrices, ??
}
\item{\code{G_obs_refit}}{
List of numeric matrices, ?? if \code{refit == TRUE} ??
}
\item{\code{graph}}{
List of \code{\link[igraph:make_graph]{igraph::graph}} objects, ??
}
\item{\code{rk}}{
Numeric vector, ??
}
}
\description{
Following the methodology from \insertCite{engelkeTaeb2024;textual}{graphicalExtremes},
fits an extremal graph structure with latent variables.
}
\examples{
??

}
\references{
\insertAllCited{}
}
\seealso{
Other structure estimation methods:
\code{\link{data2mpareto}()},
\code{\link{eglearn}()},
\code{\link{emst}()},
\code{\link{fit_graph_to_Theta}()}
}
\concept{structureEstimation}
