% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_param.R
\name{emp_chi}
\alias{emp_chi}
\title{Empirical estimation of extremal correlation matrix \eChi}
\usage{
emp_chi(data, p = NULL)
}
\arguments{
\item{data}{Numeric \nxd matrix, where \code{n} is the
number of observations and \code{d} is the dimension.}

\item{p}{Numeric scalar between 0 and 1 or \code{NULL}. If \code{NULL} (default),
it is assumed that the \code{data} are already on multivariate Pareto scale. Else,
\code{p} is used as the probability in \code{\link[=data2mpareto]{data2mpareto()}} to standardize the \code{data}.}
}
\value{
Numeric matrix \dxd. The matrix contains the
bivariate extremal coefficients \eqn{\chi_{ij}}, for \eqn{i, j = 1, ..., d}.
}
\description{
Estimates empirically the matrix of bivariate extremal correlation coefficients \eChi.
}
\examples{
n <- 100
d <- 4
p <- .8
Gamma <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)

set.seed(123)
my_data <- rmstable(n, "HR", d = d, par = Gamma)
emp_chi(my_data, p)
}
