% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{emp_chi_mat}
\alias{emp_chi_mat}
\title{Estimate matrix of \eqn{\chi}}
\usage{
emp_chi_mat(data, p, pot = FALSE)
}
\arguments{
\item{data}{Numeric matrix \eqn{n \times d}{n x d}. A data matrix
containing \eqn{d} variables.}

\item{p}{Numeric, between 0 and 1. It is the probability threshold used to
compute the \eqn{\chi} coefficient.}

\item{pot}{Boolean. if TRUE, then pot-type estimation of EC is used.
By default, \code{pot = FALSE}.}
}
\value{
Numeric matrix \eqn{d\times d}{d x d}. The matrix containing the
bivariate extremal coefficientes \eqn{\chi_{ij}}, for \eqn{i, j = 1, ..., d}.
}
\description{
Estimates empirically the extremal \eqn{\chi} correlation coefficient
given the dataset \code{data}.
}
