% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_param_MLE.R
\name{fillFixedParams}
\alias{fillFixedParams}
\title{Helper function to combine par with fixed params (in init)}
\usage{
fillFixedParams(par, init, fixParams)
}
\arguments{
\item{par}{Numeric vector. The parameters that are optimized}

\item{init}{Numeric vector. The initial parameters (including the ones optimized over)}

\item{fixParams}{Numeric or logical vector. Positions of fixed parameters in the full parameter vector.}
}
\description{
Helper function to combine par with fixed params (in init)
}
