% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flights delay data}
\format{
A named \code{list} with three entries:
\describe{
\item{\code{airports}}{A \code{data.frame} containing information about US airports}
\item{\code{delays}}{Daily aggregated delays at US airports}
\item{\code{flightCounts}}{An array containing yearly number of flights between US airports}
}
}
\source{
Raw delays data:
\itemize{
\item \url{https://www.bts.dot.gov/browse-statistical-products-and-data/bts-publications/airline-service-quality-performance-234-time}
}

Fields/Forms used in the raw data:
\itemize{
\item \url{https://esubmit.rita.dot.gov/ViewReports.aspx}
\item \url{https://esubmit.rita.dot.gov/On-Time-Form1.aspx}
\item \url{https://esubmit.rita.dot.gov/On-Time-Form3A.aspx}
}

Airports:
\itemize{
\item \url{https://openflights.org/data.html}
}
}
\usage{
flights
}
\description{
A dataset containing daily total delays of major airlines in the USA.
TODO:
\itemize{
\item add licenses
\item add interface functions?
\itemize{
\item nFlights -> connection-list?
}
\item add plot function
}
}
\details{
\code{flightCounts} is a three-dimensional array, containing the number of flights in the dataset
between each pair of airports, aggregated on a yearly basis.
Each entry is the total number of flights between the departure airport (row)
and destination airport (column) in a given year (dimension 3).

\code{delays} is a three-dimensional array containing daily total positive delays,
in minutes, of incoming and outgoing flights respectively.
Each column corresponds to an airport in the dataset and each row corresponds
to a day. The third dimension has two entries \code{'arrival'} containing delays of
incoming flights and \code{'departure'} containing delays of outgoing flights.

\code{airports} is a data frame containing the following information about a number of US airports:
\describe{
\item{\code{IATA}}{3-letter IATA code}
\item{\code{Name}}{name of the airport}
\item{\code{City}}{main city served by the airport}
\item{\code{Country}}{country or territory where the airport is located (mostly \code{"United States"})}
\item{\code{ICAO}}{4-letter ICAO code}
\item{\code{Latitude}}{latitude of the airport, in decimal degrees}
\item{\code{Longitude}}{longitude of the airport, in decimal degrees}
\item{\code{Altitude}}{altitude of the airport, in feet}
\item{\code{Timezone}}{timezone of the airport, in hours offset from UTC}
\item{\code{DST}}{
Daylight savings time used at the airport.
Either 'A' (US/Canada), 'N' (None) or NA.
}
\item{\code{Timezone2}}{name of the timezone of the airport}
}
}
\examples{
# Get total number of flights in the dataset:
totalFlightCounts <- apply(flights$flightCounts, c(1,2), sum)

# Get number of flights for specific years in the dataset:
flightCounts_08_09 <- apply(flights$flightCounts[,,c('2008', '2009')], c(1,2), sum)

# Get total delays (arriving + departing):
totalDelays <- apply(flights$delays, c(1,2), sum)

}
\keyword{datasets}
