% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flights}
\alias{flights}
\title{Flights delay data}
\format{
A named \code{list} with four entries:

\code{flightCounts} is a three dimensional array, containing the number of flights in the dataset
between each pair of airports, aggregated on a yearly basis.
Each entry is the total number of flights between the departure airport (row)
and destination airport (column) in a given year (dimension 3).

\code{arrivalDelays} and \code{departureDelays} are integer matrices containing the
total positive delays, in minutes, of incoming and outgoing flights respectively.
Each column corresponds to an airport in the dataset and each row corresponds
to a day.

\code{airports} is a \code{data.frame} containing the following information about a number of US airports:
\describe{
\item{\code{IATA}}{3-letter IATA code}
\item{\code{Name}}{name of the airport}
\item{\code{City}}{main city served by the airport}
\item{\code{Country}}{country or territory where the airport is located (mostly "United States")}
\item{\code{ICAO}}{4-letter ICAO code}
\item{\code{Latitude}}{latitude of the airport, in decimal degrees}
\item{\code{Longitude}}{longitude of the airport, in decimal degrees}
\item{\code{Altitude}}{altitude of the airport, in feet}
\item{\code{Timezone}}{timezone of the airport, in hours offset from UTC}
\item{\code{DST}}{
Daylight savings time used at the airport.
Either 'A' (US/Canada), 'N' (None) or NA.
}
\item{\code{Timezone2}}{name of the timezone of the airport}
}
}
\source{
Raw delays data:
\itemize{
\item \url{https://www.bts.dot.gov/browse-statistical-products-and-data/bts-publications/airline-service-quality-performance-234-time}
}

Fields/Forms used in the raw data:
\itemize{
\item \url{https://esubmit.rita.dot.gov/ViewReports.aspx}
\item \url{https://esubmit.rita.dot.gov/On-Time-Form1.aspx}
\item \url{https://esubmit.rita.dot.gov/On-Time-Form3A.aspx}
}

Airports:
\itemize{
\item \url{https://openflights.org/data.html}
}
}
\usage{
flights
}
\description{
TODO:
\itemize{
\item add licenses
\item convert Timezone to numeric
\item in DST: "U" -> NA
\item see R-code in paper-project
}
}
\details{
A dataset containing daily total delays of major airlines in the USA.
}
\keyword{datasets}
