% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_estimation_functions.R
\name{fmpareto_graph_HR}
\alias{fmpareto_graph_HR}
\title{Maximum likelihood parameters of a HR block graph.}
\usage{
fmpareto_graph_HR(graph, data, p = NULL, cens = FALSE,
  edges_to_add = NULL)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.
An undirected block graph, i.e., a decomposable, connected
graph where the minimal separators of the cliques have size at most one.}

\item{data}{Numeric matrix \eqn{n\times d}{n x d}. A dataset containing
\eqn{n} observations following a \eqn{d}-variate HR Pareto distribution.}

\item{p}{Numeric between 0 and 1. Threshold probability. If \code{NULL},
it is assumed that \code{data} is already distributed as multivariate Pareto.}

\item{cens}{Boolean. If TRUE, then censored log-likelihood is computed.
By default, \code{cens = FALSE}.}

\item{edges_to_add}{Numeric matrix \eqn{m\times 2}{m x 2}, where \eqn{m} is
the number of edges that are tried to be added in the forward selection.
By default, \code{edges_to_add = NULL}.}
}
\value{
List. The list is made of:
\itemize{
\item \code{graph} --- Graph object from \code{igraph} package. It represents
the block graph passed to the function.
\item \code{Gamma} --- Numeric matrix \eqn{d\times d}{d x d}. It represents
the estimated variogram matrix \eqn{\Gamma}.
}
}
\description{
Estimates the parameters of a HR block graph by maximizing the
(censored) log-likelihood.
}
