% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Gamma2chi}
\alias{Gamma2chi}
\title{Transformation of the Huesler--Reiss variogram \eqn{\Gamma} to extremal correlation \eqn{\chi}}
\usage{
Gamma2chi(Gamma)
}
\arguments{
\item{Gamma}{Numeric or matrix, with positive entries.}
}
\value{
Numeric or matrix. The extremal correlation coefficient.
}
\description{
Transforms the \code{Gamma} matrix from the definition of a Huesler--Reiss
distribution into the corresponding extremal correlation \eqn{\chi}.
}
