% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Gamma2Graph}
\alias{Gamma2Graph}
\title{Gamma2Graph}
\usage{
Gamma2Graph(Gamma, to_plot = T)
}
\arguments{
\item{Gamma}{Numeric matrix representing a \eqn{d \times d}{d x d}
variogram.}

\item{to_plot}{Boolean. If \code{TRUE} (default), it plots the produced
graph.}
}
\value{
Graph object from \code{igraph} package. An undirected graph.
}
\description{
Transforms Gamma matrix to an \code{igraph} object plots it (optionally).
}
