% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations.R
\name{Gamma2par}
\alias{Gamma2par}
\title{Extract upper triangular part of \eGamma}
\usage{
Gamma2par(Gamma)
}
\arguments{
\item{Gamma}{Numeric \dxd variogram matrix.}
}
\value{
Numeric vector with \code{d} elements.
The upper triangular part of the given \code{Gamma} matrix.
}
\description{
This function returns a vector containing the upper triangular part
of the matrix \code{Gamma}. If \code{Gamma} is already a vector, it returns
it as it is.
}
\keyword{internal}
