% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Gamma2Theta}
\alias{Gamma2Theta}
\title{Transformation of \eqn{\Gamma} matrix to \eqn{\Theta} matrix}
\usage{
Gamma2Theta(Gamma)
}
\arguments{
\item{Gamma}{Numeric \eqn{d \times d}{d x d} variogram matrix.}
}
\value{
Numeric \eqn{\Sigma^{(k)}}{\Sigma^(k)} matrix of size \eqn{(d - 1) \times (d - 1)}{(d - 1) x (d - 1)} if
\code{full = FALSE}, and of size \eqn{d \times d}{d x d} if \code{full = TRUE}.
}
\description{
Transforms the \code{Gamma} matrix from the definition of a
Huesler--Reiss distribution to the corresponding \eqn{\Gamma} matrix.
}
\details{
Every \eqn{d \times d}{d x d} \code{Gamma} matrix in the definition of a
Huesler--Reiss distribution can be transformed into a
\eqn{d \times d}{d x d} \eqn{\Theta} matrix, which
contains the graph structure corresponding to the Huesler--Reiss distribution
with parameter matrix \code{Gamma}.
}
\examples{
Gamma <- cbind(
  c(0, 1.5, 1.5, 2),
  c(1.5, 0, 2, 1.5),
  c(1.5, 2, 0, 1.5),
  c(2, 1.5, 1.5, 0)
)
Gamma2Sigma(Gamma, k = 1, full = FALSE)
}
\references{
\insertAllCited{}
}
