% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_examples.R
\name{generate_random_model}
\alias{generate_random_model}
\title{Generate random Hüsler--Reiss Models}
\usage{
generate_random_model(d, graph_type = "tree", ...)
}
\arguments{
\item{d}{Number of vertices in the graph}

\item{graph_type}{\code{"tree"}, \code{"decomposable"}, or \code{"general"}}
}
\description{
Generates a random connected graph and Gamma matrix with conditional independence
structure corresponding to that graph.
}
