% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_small_tol}
\alias{get_small_tol}
\alias{get_large_tol}
\title{Tolerances to be used in computations}
\usage{
get_small_tol(overwrite = NULL)

get_large_tol(overwrite = NULL)
}
\arguments{
\item{overwrite}{\code{NULL} or numeric scalar. If specified, use this value
instead of the option value.}
}
\value{
A non-negative numerical scalar
}
\description{
Helper function that returns the tolerance to be used in internal computations.
}
\details{
There are two different tolerances used in the package, for details see
\code{\link{graphicalExtremes-package}}. The default values for these tolerances can be
set using the options \code{"graphicalExtremes.tol.small"} and
\code{"graphicalExtremes.tol.large"}.
}
\seealso{
\code{\link{graphicalExtremes-package}}

Other default parameters:
\code{\link{get_alert_function}()},
\code{\link{get_mc_cores}()}
}
\concept{getOptions}
