% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_tol}
\alias{get_tol}
\title{Tolerance to be used instead of machine precision}
\usage{
get_tol(overwrite = NULL)
}
\arguments{
\item{overwrite}{Use this value if specified}
}
\value{
A non-negative numerical scalar
}
\description{
Helper function that returns the tolerance to be used e.g. when checking
matrices for symmetry and definiteness.
Can be set using \code{setOption('graphicalExtremes.tol', ...)}.
}
\seealso{
graphicalExtremes-package
}
