% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalExtremes.R
\docType{package}
\name{graphicalExtremes}
\alias{graphicalExtremes}
\alias{graphicalExtremes-package}
\title{graphicalExtremes: Statistical methodology for graphical extreme value models.}
\description{
The \code{graphicalExtremes} package provides three categories of functions:
simulation, estimation and transformation.
}
\section{Simulation functions}{

\itemize{
\item \link{rmpareto}
\item \link{rmpareto_tree}
\item \link{rmstable}
\item \link{rmstable_tree}
}
}

\section{Estimation functions}{

\itemize{
\item \link{fmpareto_graph_HR}
\item \link{mst_HR}
\item \link{emp_chi_multdim}
\item \link{emp_chi}
}
}

\section{Transformation functions}{

\itemize{
\item \link{Gamma2graph}
\item \link{Gamma2Sigma}
\item \link{Sigma2Gamma}
\item \link{Gamma2chi}
\item \link{chi2Gamma}
\item \link{complete_Gamma}
\item \link{data2mpareto}
}
}

\references{
\insertRef{eng2019}{graphicalExtremes}
}
