% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicalExtremes.R
\docType{package}
\name{graphicalExtremes}
\alias{graphicalExtremes}
\alias{graphicalExtremes-package}
\title{graphicalExtremes: Statistical methodology for graphical extreme value models.}
\description{
The \code{graphicalExtremes} package provides three categories of functions:
simulation, estimation and transformation.
}
\section{Simulation functions}{

\itemize{
\item \code{\link{rmpareto}}
\item \code{\link{rmpareto_tree}}
\item \code{\link{rmstable}}
\item \code{\link{rmstable_tree}}
}
}

\section{Estimation functions}{

\itemize{
\item \code{\link{fmpareto_graph_HR}}
\item \code{\link{mst_HR}}
\item \code{\link{emp_chi_multdim}}
\item \code{\link{emp_chi}}
}
}

\section{Transformation functions}{

\itemize{
\item \code{\link{Gamma2graph}}
\item \code{\link{Gamma2Sigma}}
\item \code{\link{Sigma2Gamma}}
\item \code{\link{Gamma2chi}}
\item \code{\link{chi2Gamma}}
\item \code{\link{complete_Gamma}}
\item \code{\link{data2mpareto}}
}
}

\references{
\insertRef{eng2019}{graphicalExtremes}
}
