% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/danube_plotting.R
\name{plotDanube}
\alias{plotDanube}
\title{Plot Danube River Flow Data}
\usage{
plotDanube(
  stationIndices = NULL,
  graph = NULL,
  plotStations = TRUE,
  plotConnections = TRUE,
  labelStations = FALSE,
  returnGGPlot = FALSE,
  useStationVolume = FALSE,
  useConnectionVolume = FALSE,
  mapCountries = c("Germany"),
  vertexColors = NULL,
  vertexShapes = NULL,
  edgeColors = NULL,
  xyRatio = NULL,
  clipMap = FALSE,
  useLatex = FALSE,
  edgeAlpha = 0.2
)
}
\description{
Plot Danube River Flow Data
}
\examples{
# Basic plot
graphicalExtremes::plotDanube()

# Plot flow volumes
graphicalExtremes::plotDanube(
    clipMap = 1.2,
    useConnectionVolume = TRUE,
    useStationVolume = TRUE,
    returnGGPlot = TRUE
) + ggplot2::theme(legend.position = "none")

# Plot other graph structures
nStations <- nrow(graphicalExtremes::danube$info)
g <- igraph::erdos.renyi.game(nStations, nStations, 'gnm')
graphicalExtremes::plotDanube(
    clipMap = 1.2,
    graph = g
)

}
