% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rmpareto}
\alias{rmpareto}
\title{Simulate samples of multivariate Pareto distribution}
\usage{
rmpareto(model, d, no.simu = 1, par)
}
\arguments{
\item{model}{String. The parametric model type. Is one of:
\itemize{
\item \code{HR}
\item \code{logistic}
\item \code{neglogistic}
\item \code{dirichlet}.
}}

\item{d}{Positive integer. Dimension of the multivariate Pareto distribution.}

\item{no.simu}{Positive integer. Number of simulations (by default equal to 1).}

\item{par}{Is the respective parameter for the given \code{model}.
Is one of:
\itemize{
\item \eqn{\theta \in (0, 1)}{0 < \theta < 1}, if \code{model = logistic}
\item \eqn{\theta > 0}, if \code{model = neglogistic}
\item \eqn{\alpha}, numeric vector of size \code{d}, if \code{model = dirichlet}
\item \eqn{\Gamma}, numeric matrix representing a \eqn{d \times d}{d x d} variogram,
if \code{model = HR}.
}}
}
\value{
List. The list is made of:
\itemize{
\item \code{res} Numeric matrix of size \eqn{no.simu \times d}{no.simu x d}.
The simulated multivariate Pareto data.
\item \code{counter} Positive integer. The number of times needed to sweep of
the \code{d} variables to simulate \code{no.simul} multivariate observations.
}
}
\description{
Simulates exact samples of multivariate Pareto distributions
}
