% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rmpareto_tree}
\alias{rmpareto_tree}
\title{Simulate samples of multivariate Pareto distribution from a tree}
\usage{
rmpareto_tree(n, model = c("HR", "logistic", "dirichlet")[1], tree, par)
}
\arguments{
\item{n}{Positive integer. Number of simulations.}

\item{model}{String. The parametric model type. Is one of:
\itemize{
\item \code{HR} (default),
\item \code{logistic},
\item \code{dirichlet}.
}}

\item{tree}{Graph object from \code{igraph} package.
This object represents a tree, i.e., an
undirected graph that is connected and has no cycles.}

\item{par}{Is the respective parameter for the given \code{model}.
Is one of:
\itemize{
\item \eqn{\Gamma}, numeric matrix representing a \eqn{d \times d}{d x d}
variogram, if \code{model = HR}. Alternatively, can be a vector of
length \eqn{d - 1}, containing the entries of the variogram corresponding
to the edges of the given \code{tree}.
\item \eqn{\theta \in (0, 1)}{0 < \theta < 1}, if \code{model = logistic}.
\item a matrix of size \eqn{(d - 1) \times 2}{(d - 1) x 2} containing
the relative \eqn{\alpha > 0} coefficients, if \code{model = dirichlet}.
}}
}
\value{
Numeric matrix of size \eqn{n \times d}{n x d}.
The simulated multivariate Pareto data.

## !!! add examples (define params and call function)
}
\description{
Simulates a tree graphical model, following a multivariate Pareto
distribution.
}
