% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{set_graph_parameters}
\alias{set_graph_parameters}
\title{Set graphical parameters}
\usage{
set_graph_parameters(graph)
}
\arguments{
\item{graph}{Graph object from \code{igraph} package.
An undirected block graph, i.e., a decomposable, connected
graph where the minimal separators of the cliques have size at most one.}
}
\value{
Graph object from \code{igraph} package.
}
\description{
Set graphical parameters to \code{graph} which is an object from the
\code{igraph} package.
}
