% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_transformations.R
\name{Sigma2Gamma}
\alias{Sigma2Gamma}
\title{Transformation of \eSigma and \eSigmaK matrix to \eGamma matrix}
\usage{
Sigma2Gamma(Sigma, k = NULL)
}
\arguments{
\item{Sigma}{Numeric \dxd covariance matrix \eSigma, or \d1xd1 covariance
matrix \eSigmaK from the definition of a Huesler--Reiss distribution.}

\item{k}{\code{NULL} or integer between \code{1} and \code{d}.
If \code{NULL}, \code{Sigma} is interpreted as \eSigma, else
indicates which matrix \eSigmaK is given.}
}
\value{
Numeric \dxd \eGamma matrix.
}
\description{
Transforms the \eSigmaK matrix from the definition of a
Huesler--Reiss distribution to the corresponding \eGamma matrix.
}
\examples{
Sigma1 <- rbind(
  c(1.5, 0.5, 1),
  c(0.5, 1.5, 1),
  c(1, 1, 2)
)
Sigma2Gamma(Sigma1, k = 1, full = FALSE)

}
\seealso{
Other MatrixTransformations: 
\code{\link{Gamma2Sigma}()},
\code{\link{Gamma2Theta}()},
\code{\link{Gamma2graph}()},
\code{\link{Theta2Gamma}()}
}
\concept{MatrixTransformations}
