% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_helpers.R
\name{simu_px_logistic}
\alias{simu_px_logistic}
\title{Simulates logistic extremal functions}
\usage{
simu_px_logistic(no.simu = 1, idx, N, theta)
}
\description{
Simulates logistic extremal functions
}
