% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation_functions.R
\name{Theta2Gamma}
\alias{Theta2Gamma}
\title{Transformation of \eqn{\Gamma} matrix to \eqn{\Theta} matrix}
\usage{
Theta2Gamma(Theta, k = NULL)
}
\arguments{
\item{Theta}{Numeric \eqn{d \times d}{d x d} matrix (if \code{k} is \code{NULL})
or \eqn{(d-1) \times (d-1)}{(d-1) x (d-1)} matrix (if \code{k} is a number).}

\item{k}{\code{NULL} or integer between 1 and d.
If this is \code{NULL} the input \code{Theta} is interpreted as a \eqn{d \times d}{d x d}
precision matrix \eqn{\Theta}, otherwise as \eqn{\Theta^k}.}
}
\value{
The \eqn{d \times d}{d x d} variogram matrix implied by \code{Theta}.
}
\description{
Transforms a precision matrix (\eqn{\Theta} or \eqn{\Theta^k})
to the corresponding variogram matrix.
}
\seealso{
Other MatrixTransformations: 
\code{\link{Gamma2Sigma}()},
\code{\link{Gamma2Theta}()},
\code{\link{Sigma2Gamma}()}
}
\concept{MatrixTransformations}
