% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{unif}
\alias{unif}
\title{Uniform margin}
\usage{
unif(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric vector with entries rescaled to uniform margins
}
\description{
Rescale the vector \code{x} to uniform margin.
}
